--- package copy.json
@@ -1,35 +1,0 @@
-{
-  "name": "diarybackend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon index.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/GongMolCamp/diaryBackEnd.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/GongMolCamp/diaryBackEnd/issues"
-  },
-  "homepage": "https://github.com/GongMolCamp/diaryBackEnd#readme",
-  "dependencies": {
-    "@react-native-async-storage/async-storage": "^2.1.0",
-    "axios": "^1.7.9",
-    "bcrypt": "^5.1.1",
-    "body-parser": "^1.20.3",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.7",
-    "express": "^4.21.2",
-    "firebase-admin": "^13.0.2",
-    "jsonwebtoken": "^9.0.2",
-    "mysql2": "^3.12.0",
-    "node-cron": "^3.0.3",
-    "socket.io": "^4.8.1"
-  },
-  "keywords": []
-}
