--- utils copy/socket_tool.js
@@ -1,3 +1,0 @@
-export function queryChatGroup() {
-    
-}
